extract ExploreHuggingFaceDialog to its own header
